Created a C++ backend for some calculations
Some of the functions in multipleFamilyCalculations.R are far too
inefficient when run on large data sets. Adding a C++ backend gives
better run times in the short term and allows for more optimization
work to be done in the long term. In order to prevent a barrier for
future devlopment, the original R backend has been left in and now
there is an additional parameter for choosing which backend to run
the functions with (defaulting to C++). This will still allow easy
prototyping of new methods in R.